Extract shared todo helpers and list view
Move the todo parsing/status logic into lib/todos.ts and the list
renderer into components/TodoList.tsx so a live progress dock can reuse
them without importing from TranscriptCard (which would be circular).
Adds summarizeTodos for the count/current-task summary the dock needs.
No behaviour change.
